feat: separate CLI and Lib modules for usage in Nodejs
