--- bin/run
@@ -1,20 +1,0 @@
-#!/usr/bin/env node
-
-const fs = require('fs')
-const path = require('path')
-const project = path.join(__dirname, '../tsconfig.json')
-const dev = fs.existsSync(project)
-const isTest = process.env.NODE_ENV === 'test'
-
-
-if (dev || isTest) {
-  require('ts-node').register({ project })
-}
-
-const cli = require(`../${dev ? 'src' : 'lib'}`).default
-
-cli(process.argv.slice(2))
-  .then()
-  .catch(console.error)
-
-
