+++ package.json
@@ -23,7 +23,9 @@
     "/lib"
   ],
   "bin": {
-    "compare-json-schemas": "./bin/run"
+    "compare-json-schemas": "./lib/cli/run",
+    "compare-yaml-schemas": "./lib/cli/run",
+    "compare-schemas": "./lib/cli/run"
   },
   "scripts": {
     "build": "npm-run-all clean lint build:typescript",
