+++ src/cli/run.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import main from './cli'
+
+main(process.argv.slice(2)).then().catch(console.error)
